feat: 修复实时对话和视频生成的bug。允许云端训练

+++ GaussianTalker/data_utils/deepspeech_features/deepspeech_features.py
@@ -141,7 +141,7 @@ def pure_conv_audio_to_deepspeech(audio,
     target_sample_rate = 16000
     if audio_sample_rate != target_sample_rate:
         resampled_audio = resampy.resample(
-            x=audio.astype(np.float),
+            x=audio.astype(np.float32),
             sr_orig=audio_sample_rate,
             sr_new=target_sample_rate)
     else:
